refactor(workspace): gather the mount story into one src/mount module
The "what is a mount and how does it come online" story was smeared across
config/ (loader + scaffold), workspace-apps/ (open + errors + auth-client),
and daemon/ (the Mount type, MountContext, defineMount, the name rule), while
the Mount type lived apart from its runtime witness (isMount) and its name
rule. Those splits were artifacts of the Mount[] registry era; with one config
= one mount they no longer earn three directories.

- new src/mount/contract.ts is the single, drift-proof mount contract: Mount,
  MountContext, defineMount, the DaemonRuntime/StartedMount it produces, the
  isMount guard, and the isValidMountName/validateMountNames name rule, all
  browser-safe (types + pure functions)
- config/ and workspace-apps/ collapse into src/mount/ (config-source,
  load-config, open, open-errors, auth-client); both directories are deleted
- daemon/types.ts keeps only the narrowed served view (DaemonServedMount); the
  mount-contract types move to mount/, so the mount -> daemon type edge is gone
  and mount/ is a clean leaf the daemon builds on

Pure relocation: every barrel re-exports the same symbols from the new paths,
so @epicenter/workspace, /node, and /daemon are byte-identical to consumers
and the CLI is untouched. Triple-publishing defineMount across three barrels
is left as a separate, consumer-facing cleanup.
