+++ packages/workspace/src/daemon/index.ts
@@ -1,11 +1,11 @@
+export type { DaemonRuntime, StartedMount } from '../mount/contract.js';
+export {
+	defineMount,
+	type Mount,
+	type MountContext,
+} from '../mount/contract.js';
 export {
 	type AttachProjectInfrastructureOptions,
 	attachProjectInfrastructure,
 	type ProjectInfrastructure,
 } from './attach-project-infrastructure.js';
-export {
-	defineMount,
-	type Mount,
-	type MountContext,
-} from './define-mount.js';
-export type { DaemonRuntime, StartedMount } from './types.js';
