+++ packages/workspace/src/mount/open-errors.ts
